Refactored fetchArticle into axios
